Changed: Service Worker now use network first approach
This is to prevent having an outdated collection that will lead to
request
outdated entries urls that might not exist anymore (and that have not
been
cached yet, see [#451](https://github.com/MoOx/phenomic/issues/451)).
This is for the better.
By the way, we now rely on
[offline-plugin](https://github.com/NekR/offline-plugin)
to generate Service Worker and Appcache files.
The part that register the Service Worker is now included in the
client
bundle.

Closes #485
Closes #398